25. Hunt for integer division

+++ src/Device/includes/ReSTIR/DI/SpatialMISWeight.h
@@ -233,7 +233,7 @@ struct ReSTIRDISpatialResamplingMISWeight<RESTIR_DI_BIAS_CORRECTION_PAIRWISE_MIS
 				// (Eq. 7.6 of "A Gentle Introduction to ReSTIR")
 				//
 				// We also only want that division when not using confidence weights
-				return mi / (valid_neighbors_count + 1);
+				return mi / (valid_neighbors_count + 1.0f);
 		}
 		else
 		{
@@ -251,11 +251,11 @@ struct ReSTIRDISpatialResamplingMISWeight<RESTIR_DI_BIAS_CORRECTION_PAIRWISE_MIS
 				// !!! This assumes that the center pixel is resampled last (which it is in this ReSTIR implementation) !!!
 
 				if (render_data.render_settings.restir_di_settings.use_confidence_weights)
-					return mc + static_cast<float>(center_pixel_reservoir.M) / (center_pixel_reservoir.M + valid_neighbors_M_sum);
+					return mc + static_cast<float>(center_pixel_reservoir.M) / (float)(center_pixel_reservoir.M + valid_neighbors_M_sum);
 				else
 					// In the defensive formulation, we want to divide by M, not M-1.
 					// (Eq. 7.6 of "A Gentle Introduction to ReSTIR") so 'valid_neighbors_count + 1'
-					return (1 + mc) / (valid_neighbors_count + 1);
+					return (1 + mc) / (valid_neighbors_count + 1.0f);
 			}
 		}
 	}
